+++ data/ble/ble_benchmark/create_flow_forecast_file.py
@@ -4,12 +4,16 @@ import argparse
 import os
 
 import fiona
+import fiona._env
+import fiona.env
 import geopandas as gpd
 import numpy as np
 import pandas as pd
+import rasterio._env
+from osgeo import gdal
 
 
-gpd.options.io_engine = "pyogrio"
+# gpd.options.io_engine = "pyogrio"
 
 
 def create_flow_forecast_file(
@@ -55,6 +59,8 @@ def create_flow_forecast_file(
 
     '''
 
+    print(" ******************************************")
+
     def fill_missing_flows(forecast: pd.DataFrame, nwm_river_layer: pd.DataFrame):
         """
         This function will fill in missing flows in the forecast dataframe.
@@ -81,7 +87,7 @@ def create_flow_forecast_file(
 
             nonintersects = merged[merged['discharge'].isna()]
 
-            print(f'Iteration {n}. {len(nonintersects)} segments remaining.')
+            # print(f'Iteration {n}. {len(nonintersects)} segments remaining.')
 
             updated = False
             for i, row in nonintersects.iterrows():
@@ -137,7 +143,7 @@ def create_flow_forecast_file(
     xs_layer = gpd.read_file(ble_geodatabase, layer=ble_xs_layer_name)
 
     # Read in the NWM stream layer into a geopandas dataframe using the bounding box option based on the extents of the BLE XS layer.
-    nwm_river_layer = gpd.read_file(nwm_geodatabase, bbox=xs_layer, layer=nwm_stream_layer_name)
+    nwm_river_layer = gpd.read_file(nwm_geodatabase, mask=xs_layer, layer=nwm_stream_layer_name)
 
     # Make sure xs_layer is in same projection as nwm_river_layer.
     xs_layer_proj = xs_layer.to_crs(nwm_river_layer.crs)
@@ -172,7 +178,6 @@ def create_flow_forecast_file(
         forecast['discharge'] = forecast['discharge'] * dischargeMultiplier
 
         # Assign flow to segments missing flows
-        print(f'Filling in missing flows for {return_period[i]} flow.')
         forecast = fill_missing_flows(forecast, nwm_river_layer)
 
         # Set paths and write file
@@ -181,6 +186,8 @@ def create_flow_forecast_file(
         os.makedirs(dir_of_csv, exist_ok=True)
         path_to_csv = os.path.join(dir_of_csv, "ble_huc_{}_flows_{}.csv".format(huc, return_period[i]))
         forecast.to_csv(path_to_csv, index=False)
+
+    return
 
 
 if __name__ == '__main__':